Move slow tests related fixtures in tests_slow/conftest.py
- slow_tests_datadir
- data_path
- baseline
